Move interactive commands to utils file
