+++ src/commands/kv/mod.rs
@@ -16,11 +16,6 @@ pub mod bulk;
 pub mod key;
 pub mod namespace;
 
-// Truncate all "yes", "no" responses for interactive delete prompt to just "y" or "n".
-const INTERACTIVE_RESPONSE_LEN: usize = 1;
-const YES: &str = "y";
-const NO: &str = "n";
-
 // Create a special API client that has a longer timeout than usual, given that KV operations
 // can be lengthy if payloads are large.
 fn api_client(user: &GlobalUser) -> Result<HttpApiClient, failure::Error> {
@@ -118,22 +113,6 @@ pub fn get_namespace_id(target: &Target, binding: &str) -> Result<String, failur
         binding,
         target.name
     )
-}
-
-// For interactively handling deletes (and discouraging accidental deletes).
-// Input like "yes", "Yes", "no", "No" will be accepted, thanks to the whitespace-stripping
-// and lowercasing logic below.
-pub fn interactive_delete(prompt_string: &str) -> Result<bool, failure::Error> {
-    println!("{} [y/n]", prompt_string);
-    let mut response: String = read!("{}\n");
-    response = response.split_whitespace().collect(); // remove whitespace
-    response.make_ascii_lowercase(); // ensure response is all lowercase
-    response.truncate(INTERACTIVE_RESPONSE_LEN); // at this point, all valid input will be "y" or "n"
-    match response.as_ref() {
-        YES => Ok(true),
-        NO => Ok(false),
-        _ => failure::bail!("Response must either be \"y\" for yes or \"n\" for no"),
-    }
 }
 
 fn url_encode_key(key: &str) -> String {
