+++ src/commands/kv/namespace/delete.rs
@@ -5,12 +5,13 @@ use crate::commands::kv;
 use crate::settings::global_user::GlobalUser;
 use crate::settings::toml::Target;
 use crate::terminal::message;
+use crate::terminal::utils;
 
 pub fn delete(target: &Target, user: &GlobalUser, id: &str) -> Result<(), failure::Error> {
     kv::validate_target(target)?;
     let client = kv::api_client(user)?;
 
-    match kv::interactive_delete(&format!(
+    match utils::interactive_delete(&format!(
         "Are you sure you want to delete namespace {}?",
         id
     )) {
