+++ src/terminal/mod.rs
@@ -1,2 +1,3 @@
 pub mod emoji;
 pub mod message;
+pub mod utils;
