+++ src/terminal/utils.rs
@@ -1,0 +1,26 @@
+// For interactively handling  reading in a string
+pub fn interactive_get_string(prompt_string: &str) -> String {
+    println!("{}", prompt_string);
+    let input: String = read!("{}\n");
+    input
+}
+
+// Truncate all "yes", "no" responses for interactive delete prompt to just "y" or "n".
+const INTERACTIVE_RESPONSE_LEN: usize = 1;
+const YES: &str = "y";
+const NO: &str = "n";
+// For interactively handling deletes (and discouraging accidental deletes).
+// Input like "yes", "Yes", "no", "No" will be accepted, thanks to the whitespace-stripping
+// and lowercasing logic below.
+pub fn interactive_delete(prompt_string: &str) -> Result<bool, failure::Error> {
+    println!("{} [y/n]", prompt_string);
+    let mut response: String = read!("{}\n");
+    response = response.split_whitespace().collect(); // remove whitespace
+    response.make_ascii_lowercase(); // ensure response is all lowercase
+    response.truncate(INTERACTIVE_RESPONSE_LEN); // at this point, all valid input will be "y" or "n"
+    match response.as_ref() {
+        YES => Ok(true),
+        NO => Ok(false),
+        _ => failure::bail!("Response must either be \"y\" for yes or \"n\" for no"),
+    }
+}
